test(logic): split incremental validator equivalence corpus for size gate
Split the #2237 equivalence corpus into focused files with shared helpers so each changed logic test file stays under the CI 400-line size gate while preserving coverage.
